spring-guides/gs-multi-module/application/src/main/java/io/lenur/multimodule/controller/ApplicationController.java
package io.lenur.multimodule.controller;

import io.lenur.multimodule.service.ApplicationService;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class ApplicationController {
    private final ApplicationService applicationService;

    public ApplicationController(ApplicationService applicationService) {
        this.applicationService = applicationService;
    }

    @GetMapping("/")
    public String home() {
        return applicationService.message();
    }
}
